[FIX] fixed ambiguous function calls with casts and removed C11 code

+++ src/openms/source/ANALYSIS/XLMS/XQuestScores.cpp
@@ -106,11 +106,11 @@ namespace OpenMS
 
     if (is_xlink_spectrum)
     {
-      a_priori_p = (1 - ( pow( (1 - 2 * tolerance_Th / (0.5 * range)),  (theo_size / n_charges))));
+      a_priori_p = (1 - ( pow( (1 - 2 * tolerance_Th / (0.5 * range)),  (static_cast<double>(theo_size) / static_cast<double>(n_charges)))));
     }
     else
     {
-      a_priori_p = (1 - ( pow( (1 - 2 * tolerance_Th / (0.5 * range)),  (theo_size))));
+      a_priori_p = (1 - ( pow( (1 - 2 * tolerance_Th / (0.5 * range)),  static_cast<int>(theo_size))));
     }
 
     double match_odds = 0;
